Static branch fixups (#386)
Assorted changes in support of static builds:
 . added pytext fixture for link mode
 . static and shared builds are now performed in separate subtrees
 . Test files updated to build and run with the new paths
 . Examples update to build with the new path and run with CWD=source file location directory
 . CI/CD updated to perform static builds of examples on non-MacOS platforms

[ committed by @billschereriii ]
[ reviewed by @ashao  ]

+++ examples/serial/c/example_put_unpack_1D.c
@@ -44,7 +44,7 @@ int main(int argc, char* argv[]) {
   float* tensor = (float*)malloc(dims[0]*sizeof(float));
 
   for(size_t i=0; i<dims[0]; i++)
-    tensor[i] = ((float)rand())/RAND_MAX;
+    tensor[i] = ((float)rand())/(float)RAND_MAX;
 
   void* client = NULL;
   bool cluster_mode = true; // Set to false if not using a clustered database
